Learning about buffers
This time from Sol's Graphics Tutorial

https://solhsa.com/gp2/index.html

A lot of fun.

At the moment there's quite a mess.
Though, aside from that, a buffer is set up, and updating an animal is a matter of writing to the buffer.
As this mutates the buffer, and as there's no other routine, this creates a trail.