Sprites, and PNGs.
Not too much of interest, though much of work.

Sprite(s) can now be read from PNGs, using libpng.
And, these are now held in a custom class, abstracting over size, etc. as much as possible.

As PNGs are used the CMake setup now handles a resource folder.
And, whereami is used to resolve where this folder is after execuation (relative to the binary directory).

Some work still to be done on correctly writing sprite data, though.