Maze parsing
To allow for storing of maze data as a resource this commit introduces a simple parser for mazes stored in plain text.
The parser is DIMACS-style, using a preamble for information, and line prefixes to guide parsing.
Unlike DIMACS line prefixes are used for all items.
The key aspect are:
- uint width / height values are required in the preamble, with 'w'  and 'h' prefixes, respectively.
- lines of the maze use the 'm' prefix, and if the line is shorter than the expected width, spaces are assumed.

Elsewhere, one of my std::filesystem::path misunderstandings became clear, and has been fixed.

Otherwise, some includes have been tidied...