Logic abstraction
The idea is to have a solver for each anima.
And, a term manager can be shared between solvers.
As each solver will represent, mostly, of the same thing, the idea is to create a unique term manager and store shared terms (and sorts, etc.) in a library struct.

This commit sketches the above and adds the sorts defined so far to the library.

Also, the deduction.c file is simplified further.