The simple solve, in smt2 form
Expectation is to keep things in smt2 form.
Hopefully.
As, then, the emgine can guarantee the assertion of some relevant predicates, and, otherwise, deduction can be edited without going through the source.

The smt2 sketch differs from the source in the use of quantifiers.
I'd like to have this option, and with smt2 things are much easier to edit if it turns out this leads to issues later on.