Revise term construction with parsing
And, in time, reading from files.

Parsing is mostly simple and is limited to initial setup.
Solves in the core loop continue to use built terms.
However, there is not clear link between parsing and term creation.
So, terms used within a solve are cached within a AnimaTerms struct.
As it's only possible to get terms and commands when parsing, this limits what can be made without use of the parser.
Though, I expect terms to be sufficient.