Unified symbol delcaration
With parsing this is a little messy, though I think mostly safe.

The constraint is that parsing requires a solver.
This makes sense, as arbitrary SMT statements may be parsed.

However, I want a global term and symbol managers.
For this, parsing with respect to any specific solver is no good.

So, the idea is to set up a temporary solver.
With this, parsing completes as expected.
And, so long as onl symbol information is parsed, I expect all relevant information is held in the term / symbol manager.

In testng, this seems to be confirmed.

In addition to the above, a constant for the logic used is set.
In the previous commit this was updated to UFLIA without comment.
The motivation is to use ints for spatial (etc.) reasoning.