Exploration into threads
At the moment, broken.

Though, some things:

- A bug in rendering with kTileSize not being set is fixed.
  This solves the unexpected observed rendering behaviour when a spite is not creater on the same thread as the renderer.
  And, more generally, significant problems.

For threads, I had assumed thread local solvers would be ok.
So, things are set up so that animas are created on main before being passed over to a thread.
And, once on the thread a mind is constructed.
Then, in the game loop the thread has a sleep / wake cycle to ensure a solve happens at least every so often, but long solves may run until completion.

With a single anima this works fine.
With multiple, things break.

I read https://github.com/cvc5/cvc5/discussions/10265 a little to optimistically, maybe (as being about inter-thread actions, which shouldn't happen with the above...).
Or, maybe there is some other issue...