Mutex improvements
the suspend flag and mutex, along with the resume condition are now anima local.
And, flags are now atomic.

And, with various other changes for this, the main thread no longer waits to aquire a mutex.

I expect some other things remain overlooked, though this is taking shape a little.