Progress toward pixels
Before this commit, most of the game logic worked on the bases of a 28x31 grid.
It's a little more interesting, I expect, to run the logic using pixel based locations.
For the standard maze, not too much changes.
Though, things could very much change in a non-standard maze.

As an intermediary step this amounts to using kPIXELS to handle the pixel limits in most cases.
Though, to be generalised at some point.

And, writing out a way to scale up input grids, as this seemed easier than writing out mazes by pixel.
As a consequence, the handling of blanks is lost, though checks on the expected dimensions are a trade for this.

Collision is broken, due to the offset now being per pixel rather than per tile.
So, animas can move one tile into a wall.
The effect is quite nice, though.