Rework terms a little
Current goal is to settle on a foundation for world/mind syncing.

Roughly, the idea is that on the main thread will be a world struct which contains, roughly, booleans for the union of all minds.
This is linked to terms, and to help simplify things terms are to be collected into mutually exclusive structs, if needed.
This commit generalises the term handling, a little.