Sprite location refactor
The general organising principle for the abstract state is that each anima acts independently, but there's sufficient thread safe information to maintain some shared world.
This can be extened to rendering, and arguably should, as the same data-simplifying principle also simplifies reasoning about representation.
In addition, there's the background motive of keeping most/all rendering information on the main thread.