Update cmake w/ nix prep, and package updates
Mostly a merge of the libraries file.
However, due issues getting with getting clangd to work some commented use of PkgConfig to resolve includes is kept.

stumpless has been removed, to be replaced with something else.
With a flake things will be a little easier to configure, and for some reason a pull on stumpless always required verification, which slowed down the clean process considerably (infrequent, but an issue when debugging build issues).

A cmake module for criterion is included from:
https://github.com/Snaipe/Criterion/blob/bleeding/dev/cmake/cmake/FindCriterion.cmake