initial flake
derived, mostly, from:

https://github.com/the-nix-way/dev-templates
https://discourse.cmake.org/t/how-to-correctly-generate-compile-commands-json-for-library-dependency-using-pkgconfig/15135

in the second link, skittishdev details a setup for criterion, which works well
the remainder of the flake is some basic setup for the project

the key to getting this all to work with an LSP usin clangd is .envrc and related tools (direnv / envrc.el).

roughly, generated compile commands do not include default search paths, and some libraries such as SDL3 and libpng are found one some default path.
and, as these defaults are environment dependent, clangd sees things quite differently from inside/outside nix.
though, with direnv to help manage context, a server instance can get going with the right outlook.

the criterion situation, on the other hand, i don't have a clear grasp on, other than all dependencies are included