distractions with python, though some basics with z3
initially the idea was to pass some predicates to a (cvc5) solver
though, now that it comes to things, it'd be much more interesting to try and keep as much reasoning in the solve as possible
this isn't too straightforward
though, it does seem z3 is likely of more use here, as it provides optimisation, and various other interesting things (like tactics)
so, initial experimentation with z3 has started

of note, i sure do overuse predicates (on initial passes, at least...)
and, one idea for dealing with numeric limits is to use tiles and restrict quantification to tiles, etc.

clingo, also, remains an option, esp. for optimisation
though, it remains something i'm not too interested in, as continued exploration of smt remains a strong motivatior for this project

all the above is ok
getting a somewhat usable setup for enjoyable python work, though...
it may be ok now...