some timings, and a revised source
the revised source disambiguates paths and non-wall inacessible tiles.
and, with this the checks in code for a path now require presence of a path (as would have made more sense from the start).

with this the non-optimal solves are much clearer
and, in general these are certainly sufficiently fast

the optimal solves, though, remain far too slow
though, given how inefficient the non-optimal solves are, if optimality is by pushing down a bound, there may be some less-than-difficult progress to be made...