compiler options
somehow a mess has been made of clang

I'm very unclear on the source of the problem, though the sypthoms are simple:

Attempting to open a file (the maze file) prior to a null check returns a pointer.
In main.c, the file can be read, closed, etc. no issues.
In maze.c the issue happens, the file can be opened ok, but on a check appears to be null(ptr).

lldb is also in a confused state, so I assume this is a local issue.
And, it occurs on both 19 and 21.

gcc, though, handles things ok, it seems.

The nix flake has been updated a little to get a gcc env going.
It's a mess, but this is a very unexpected fault found while attempting to diagnose a confusing fault with the z3 api, so paitence for tidying has been lost.

gcc does not help with the z3 api issue.