failures with Z3_mk_constructor
narrowed, I think, together with the source, which is tbd

the purpose of sort_refs is a mystery, still.
using the id seemed plausible, though failed to work.
notes to revisit with.

as an aside, I think the custom datatype is worthwhile.
as, it allows positions to be the value of various functions.
of course, a u16 could also work, but then the domain isn't so clearly limited...