Resolving z3 updates
In conjunction with the previous commit, z3 is now found ok.
cmake cannot find the package, but the includes and lib and accessible.
and, the same is true for cvc5.

in addition, there's an option to build the c api tests.
given these build ok, but not within smt-man, I think it's reasonable to assume something quite bad is happening in the smt-man code, somewhere.
This, perhaps, is also the source of the clang issue.

stumpless removed as it's a real hassle to confirm a clone.