things are ok ... ?
selective cuts from the z3 c api examples led to faults.
those faults persisted even when everything else in the codebase was removed (aside from headers and the z3 link).
so, perhaps it was something with the cut, a failed replacement of some supporting fn???
cutting the entire test worked out, and so far cuts have not led to an issue.

though, this has highlighted a nice tuple example in the tests to investigate...