Maze work
Right, there was a time not so long ago where the idea was to keep things pixel based.
With this the maze was stored as pixels, and so enlarged, etc.

Aside from some mixing of uint8_t and char to no particular harm, this is ok.
Though, no longer relevant, I expect.
Instead, the approach should be to scale the maze to the size of the abstract board.
If the maze is small, things in the abstract are small, and no otherwise.

Whether to keep the pixel scaled maze, though, I'm not sure.
As, it simplifies the pixel mapping.

Well, for now, regardless, an abstract maze has been added.
With this it should be mostly simple to create the path fn in the z3 API.
Creation of the tiles is included.