Renderer issues
Unresolved, unclear on what to do.

Sometimes the renderer is hek'd.
It appears this is due to something that happens in the update routine.
Though, this isn't of too much surprise, as it's where the pixel buffer is copied over.
Working through the docs and issues hasn't led to much insight.
Changing the renderer had no effect, though no easy option to try vulkan at the moment.

I wonder, perhaops, if locking only those regions required for updating animas would be something?

In any case, a few variables are refactored, and the render routine is modified a little to hopefully help kep some memory of what (was not) learnt today.