explorations: positivity
Asserting a positive constraint (i.e. or rather than nand) on the possible path enum for non-anima tiles is of *significant* help, it seems

This seems a reasonable aide to use.
And, I expect, an oracle which asserts a stronger restriction on some tiles which must be on a shortest path could help things further.
Though, given the speed of these solvers (<.5s for distant points), I think this might be sufficient for now.

Oh, though allowing a slower solver as a different tactic would also be interesting.