z3: simplify non-empty hints
Collapsing cases, recognising scope, etc.

The structure is now to consider the relevant surrounding tile, generate the disjunction, and then assert the disjunct conditional on the tile mapping, as before.

This unifies very many of the if-then conditions for assertions at the end of the fn, while also clearly establishing the scope of the disjunctions.

At the moment each group of assertions *could* be further simplified, as there are no checks other than whether the tile is accessible.
Though, the fn is set up so that certain disjuncts could, in principle, be filtered for a tighter collection of constraints.