File issue resolved
After opening the maze file, the execution entered a while loop with the condition that the read character was not the end of file.
Storage for the read character was not initialised.
It seems as though given this, compilation assumed the storage always hit EOF, and so the loop was never seen, and the file could be closed immediately.
Sure, whatever.