Generic pairs
Based on the template header sketch by sir winesalot: https://btmc.substack.com/i/159022058/good-template-header

Primarily to allow for generic structs, but also to mess with the preprocessor a little.

Changes throughout the rest of the codebase are due to addng an extra `_` in the pair struct, to avoid uppercasing the first character of the generalised type.