Maze u8 restrictions
Restriction of maze dimension to u8s.
The kind of thing I'd do in Rust, so will do here also.
Though, requires reasoning a little more about what's happening as the compiler is less helpful.
E.g., area method on pairs removed, as for u8s it's easy to overflow.