struct for sheet offsets
simple, but after much thinking about various things (though no research) this seems reasonable.

the mapping needs to be stored somewhere.
methods seem excessive, as the relevant information is a static.
and, given the information is quite structured without too much depth (size / offsets) a struct for the relevant entries seems fine.

elsewhere, the size of a tile to be drawn has been reduced to an int, as the tiles are all squares, and this will not change.