Initialisation, and animas
Anima count bumped to 4, as there's no need to introduce a sheet for each anima.

Elsewhere, initialisation has been updated to be wrt. some area of memory.
So, instead of Anima_create(...)  returning an anima, the fn is now Anima_create(Anima *anima, ...), with the anima created at the given pointer.
There are two motivations for this:

- Memory ownership is a little clearer, typically --- memory is owned by the fn which allocates the memory.
- There return value of the initalisation fns can be used for error handling.

Most of the fns return void (for the moment), so the second motivation is yet to be realized.