Anima restructuring
Animas are now general structs containing all relevant thread local things.
In particular, then, minds are a field, and so are mutexes, etc.
In this respect, there's various redirects in this commit.

Otherwise, the key differences are with respect to minds.
The idea is that a mind alone should be sufficnent for a solve.
So:
- An animas pov/view is now a field in a mind, as this specifies how things are.
- An id is stored as a field in a mind, as solves are (typically) indexical.

With this, also, some methods have been moved from animas to minds.