Refinement
TILE_SCALE almost eliminated.
In place, two struct fields, in the renderer and animas.
Renderer, to determine the frame buffer size.
Animas, to determine when to move, etc.
In time this can be revised to a dynamically generated value determined by the maze and sprite sheet.

Elsewhere:

Renderer initialization is now through an argument, in line with other initialization.

A pallete struct. (distracted.)