Tile data
The final handful of tasks before implementing the persona.

At the moment there's a simple mapping between tiles and sprites.
In practice, sprites are twice the size of tiles.
Though, to resolve this requires a proper representation of the maze (as anima sprites appear on arbitrary maze tiles).

Handling maze tiles is something of a headache, as the ideal is to read the maze from a file.
Solves, in principle, should be compatibe with different mazes.

Mazes are built from lots of tiles, though, so mapping characters isn't so easy.
And, in any case, makes the maze input quite messy.

So, some deduction from a slightly more descriptive base seems the way to go.

maze.h contains an intial pass, which totals 16 possible permutations.
Though, it may be possible to infer the relevant edge case.