Rendering lines
I think to draw the maze it's slightly more interesting to do so directly, rather than by using sprites.

It's (reasonable to think) doing so is more efficient, as there's no need to loop over empty pixels.
Though, the primary motivation is abstraction and the possibility of compositing some simple fns, and doing more interesting things in the future (without needing to extend the sheet).

For the moment, an fn to draw a line, and maze tiles updated to be the outline of the maze.