Circles
An implementation of Bresenham's algorithm for drawing circles.
Based on the very nice writeup by Arver: https://funloop.org/post/2021-03-15-bresenham-circle-drawing-algorithm.html

I think this may look *good enough*, though is not ideal.

The rendering is split into quadrants, with the assumption that the centre point is offset up and to the right.
For the moment, at least.

Time for something else...