Initial arc work
Inferring the tiles is something of a headache.
So, to help simplify things a few passes will be made.

The first pass will check the outer edges of a maze.
As, with this pass complete +/- offsets from `within` the maze will always be safe to dereference.

At the moment a pass over the leftmost and rightmost columns is made.
These are mostly identitical, and there may be scope for abstraction.

Otherwise, the key revision is arc rendering.
Arcs now fill a tile, and are offset appropriately (I think, ...).

As the tile data was becoming a little complex with sub-unions, the data structs for particular types have been split off.