Update steps in direction
The fn has been revised to be a pair method.
Ideally it'd be a generic method, but the reliance of MAX constants makes this far more difficult than it should be.
So, for now it's included in the pairs header.
And, as this header is designed to be used with an implementation def, the method source is also ifdef'd.