Turns introduced
Rendering a sprite now takes a turn argument, where turns are one(/zero), quarter, half, or three quarter (located in the enums header).

In addition, the draw and erase methods have been refactored to highlight the relevant differences between each switch case.

Not mentioned in the last commit are a some small optimisations related to offset calculation.
In the method, repeat offsets are explicitly stored, and the fn itself caches some common values to help avoid division --- or so one would expect as the only variable is a constexpr.