Movement patterns
My initial approach to separation of actions from frames was to implement actions conditional on some number of frames passing.
This works ok, though has the drawback of the only variable being the delay.
Though, this may be sufficient for some purposes.

Still, in the base game there are varying speeds, and while close adherence isn't too important, it'd be nice to make a close approximation.
So, what else could be required?

Well, there's conflicting infomration.
The base game runs at (roughly) 60fps, this much makes sense, and seems a reasonable upper bound on movement speed --- 60pps (pixels per second).
The dossier, though, suggests this isn't the upper bound, and that sometimes objects can move at 75pps.
A slight headache, for sure, as now some frames must involve a double move.

Still, is this correct?

Some further research, and mburkley has a nice translation of the base game to C.
And, ..., and! a writeup of some movement related things.

https://pacmanc.blogspot.com/2024/05/movement-patterns.html

Abstracting a little, it seems the base determines whether to move on a frame by examining a stream of bits.
So, the maximum movement is, right, 60pps.
This seems sensbile, but the really neat part is the stream, as this allows for arbitrary move-or-not sequences.
These streams are implemented by rotating bitvecs, and so have limited variance, though it's all much more elegant than the delay I was using.

Given this, the frame and actions per frame fields have been removed.
Actions remain in order to determine sprite offsets.

Maybe the 75pps is still correct.
For now this is sufficient.