Directions as flags
To help remember a C23 feature, with no intention of use.

The underlying type of the direction enum is fixed as a u8.
And, each direction is aligned to a bit to support multiple directions.

Thinking on this a little more, it's simple to split the u8 into current and future directions.
Though, I'm not sure if it makes conceptual sense to link intent and momentum.