Local intent
Anima directional intent is now mind(/thread) local and does not appear in the situation.
This avoids some synronization, but more importantly allows for refining the sitaution to things which are observable.

This is a slight change in directive from sitautions being sufficient for a solve, if one wished to use directional intent.
Though, in general there are may things one may wish to use in a solve, and passing these through as parameters should be fine.
There should be no need to expand the situation to incorporate these so long as they have no 'observable' representation.