Abstract personas...
... and some misc. tidying.

Similar to abstract animas.
And, if the structs do not diverge they may be unified.
For the moment, though, there's a lot left to implement.

Synronization of situations is expanded to include all relevant features of the situation.
Though, animas remain unaware of other animas.
A two-pass sync may be worthwhile, to allow for the option of animas reasoning about other animas...