Persona methods, and related changes
Persona methods are mostly a copy of anima methods, nothing of too much interest.

Though, in copying things a few changes have been made...

With movement patterns there's no need for velocity.
So, velocity has been removed.
To do so, a new `DIRECTION_NONE` variant has been added to the direction enum.
Something like this is needed as some way of not moving a sprite must exist.
For the animas, no direction seems nice, as this can be used for when solving, e.g., and may have a custom sprite.

For the persona, I expect this is suffiient for handling getting stuck in corners.
As, for a persona the sprite can be drawn from direction intention, and updated from direction actual.

Otherwise, some light renaming, and some lexigraphic ordering.