A movable persona
For the moment the anima facing up is used as a sprite.

Otherwise, the persona is set up, and can be moved (without elegance).

Around the edges some work to align the persona struct and methods with animas (e.g. removing sprite_location arguments).
And in the core, very little to do.
The key parts were moving event handling to the right place, and recognising the need for bounds checking on both events and tiles.

Some work needs to go into handling the persona, but this should be fine when multiple directions are available...