Persona movement fix, part B
Common `is tile` queries generalised to a renderer method.
Renderer, as plausibly TILE_SIZE owned by the renderer, and this is the key variable.

Maze location has been updated to work with non-tile locations, by rounding each coordinate to the closest tile.

And, the persona now has an `off tile` method, which allows changes to direction so long as the direction is on the same plane.
The revision of enums to bitfields finally of use here.