Persona rendering
The basics, at least.

The persona differs a little from the base game by enlarging and centering the circle.
This is required, at the moment, as only sprite rotation is supported.
And, with an offset sprite, this can get a little messy.

Whether or not the variance will be kept, unclear.
I think I prefer the symmetry.

In any case, support for manipulating a sprite in place should be added, as this would allow for stacked transformations.

For rendering, some additional abstraction introduced in the form of an action enum, allowing for unified draw/erase methods.

And, generally, size information is now located within the persona/animas, as it's quite convenient there, and I don't anticipate interest in varying sprite sizes for these objects.

Offsets have been updated to use implicit positional initialization, to maybe help readability.
This may be reverted, though...