Rendering improvements
Before this commit, transformations in rendering were made when mapping from the sprite sheet to the frame buffer.
This limited the transformations that could be applied to sprites from the sheet.

The architecture is now:

- A pre buffer, with a size some multiple of the pixels for a tile (currently 4).
- Methods for copying sprites to the pre buffer, mutating the pre buffer, applying a pallete to the pre buffer, and writing to the frame buffer from the pre buffer.
- Sprite removal is now handled by filling the relevant pixels as transparent.

So, the general flow is: copy, mutate, write.

All the methods are parameterized to a size, so there's no overhead of excess pixels in the pre buffer.

With this, rendering is simplified significantly.
In particular, as there's no need to specify transformations as arguments.
Also a drawback in some cases, but for the moment it's certainly an improvement.

The erase method has been removed.
The draw from sheet method is unused, and could also be removed.
Though, in some cases it might be useful to avoid the overhead of the pre buffer and render directly.