Initial separation of maze reading and rendering
The idea, roughly, is to expand the tile information to support rendering without context.
This requires at least one additional pass of the data prior to writing the maze, and happens in that method.
So, it makes some sense to split then fn.

As a first pass, the split is complete, though is messy.

An additional surface fill method is introduced, and most surface mutation could/should be moved from the renderer to a surface.