Completing maze data
Completing the data is not dependent on the renderer, ... so long as the pixel count is fixed.
And, a variable pixel count should be handled by some global thing.
So, filling out maze data has been made a method on a maze struct.
And, filling line data has been extracted and commented a little.