Sprite fill borders
When erasing sprites by a fill maze information is lost.
Still, all sprites have at least a pixel border.
And, if corners are drawn as in the base game, then leaving the pixel border when performing a fill retains the maze.
So, rendering sprites has been updated with a flag for what to do with the border.

Generally speaking, the method is likely only to be used with the flag set to false (to not fill the border), and so it would be more efficient to have a fixed fn for this case.
For now, though, a flag is ok.